Extract CameraUpdate inner class
